+++ test/property_fixes/PropertyFixesTest.php
@@ -21,4 +21,19 @@ class PropertyFixesTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals(null, $obj->cancelState);
         $this->assertEquals(null, $obj->cancelStatus);
     }
+
+    /**
+     * Incorrect documentation https://developer.ebay.com/Devzone/post-order/types/ItemEligibilityResult.html
+     * Example of correct property names returned in the API https://developer.ebay.com/Devzone/post-order/post-order_v2_cancellation__post.html#Samples
+     */
+    public function testItemEligibilityResult()
+    {
+        $obj = new Sdk\PostOrder\Types\ItemEligibilityResult();
+
+        $obj->itemId = '123';
+        $this->assertInternalType('string', $obj->itemId);
+
+        $obj->transactionId = '123';
+        $this->assertInternalType('string', $obj->transactionId);
+    }
 }