+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 ### Fixes
 
-* renamed state to cancelState and status to cancelStatus.
+* Renamed state to cancelState and status to cancelStatus.
   This matches the properties returned by the API.
+* Changed type from integer to string for ItemEligibilityResult::itemId and ItemEligibilityResult::transactionId
 
 ## 11.0.0 - 2017-03-20